# src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Commune API",
    "version": "1.0.0",
    "description": "The Commune API lets AI agents and developers send and receive email, manage inboxes and domains, and process structured data from email conversations. All endpoints require a Bearer API key.\n\nBase URL: `https://api.commune.email`\n\nAuthentication: Pass your API key as `Authorization: Bearer comm_...`",
    "contact": {
      "name": "Commune Support",
      "url": "https://commune.email",
      "email": "support@commune.email"
    },
    "license": {
      "name": "Business Source License 1.1",
      "url": "https://mariadb.com/bsl11/"
    }
  },
  "servers": [
    {
      "url": "https://api.commune.email",
      "description": "Production"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "comm_...",
        "description": "API key issued from the Commune dashboard. Prefix `comm_`."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message"
          }
        }
      },
...
      "InboxWebhook": {
        "type": "object",
        "properties": {
          "endpoint": { "type": "string", "format": "uri", "description": "URL to receive webhook POSTs" },
          "events": {
            "type": "array",
            "items": { "type": "string" },
            "description": "Event types to subscribe to (default: all)"
          },
          "secret": { "type": "string", "description": "HMAC signing secret (write-only, not returned in reads)" }
        }
      },
...
      "Inbox": {
        "type": "object",
        "required": ["id", "localPart"],
        "properties": {
          "id": { "type": "string", "description": "Unique inbox ID (UUID)" },
          "localPart": { "type": "string", "description": "Local part of the inbox email address (e.g. `support` in `support@yourdomain.com`)" },
          "address": { "type": "string", "description": "Full email address, e.g. `support@yourdomain.com`" },
          "displayName": { "type": "string", "description": "Human-friendly display name" },
          "status": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "webhook": { "$ref": "#/components/schemas/InboxWebhook" },
          "extractionSchema": { "$ref": "#/components/schemas/ExtractionSchema" },
          "domain_id": { "type": "string", "description": "ID of the domain this inbox belongs to" },
          "domain_name": { "type": "string", "description": "Domain name" }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid API key",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "Unauthorized" }
          }
        }
      },
      "Forbidden": {
        "description": "Insufficient permissions",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "BadRequest": {
        "description": "Invalid request body or parameters",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "TooManyRequests": {
        "description": "Rate limit exceeded",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "Rate limit exceeded" }
          }
        }
      },
      "InternalError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      }
    }
  },
  "paths": {
...
    "/v1/inboxes": {
      "get": {
        "tags": ["Inboxes"],
        "operationId": "listInboxes",
        "summary": "List all inboxes",
        "description": "Returns all inboxes across all domains for your organization.\n\nRequires `inboxes:read` permission.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of inboxes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Inbox" }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "post": {
        "tags": ["Inboxes"],
        "operationId": "createInbox",
        "summary": "Create an inbox",
        "description": "Create a new inbox. Domain is auto-resolved if not provided — your first custom domain is used, or the shared `commune.email` domain on free plans.\n\nRequires `inboxes:write` permission.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["local_part"],
                "properties": {
                  "local_part": {
                    "type": "string",
                    "description": "Local part of the email address (e.g. `support` for `support@yourdomain.com`)"
                  },
                  "domain_id": {
                    "type": "string",
                    "description": "Domain to create the inbox on. Auto-resolved if omitted."
                  },
                  "display_name": { "type": "string", "description": "Human-friendly display name" },
                  "webhook": { "$ref": "#/components/schemas/InboxWebhook" },
                  "status": { "type": "string" }
                }
              },
              "example": {
                "local_part": "support",
                "webhook": {
                  "endpoint": "https://yourapp.com/webhooks/email"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Inbox created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/Inbox" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
    "/v1/domains/{domainId}": {
      "get": {
        "tags": ["Domains"],
        "operationId": "getDomain",
        "summary": "Get a domain",
        "description": "Returns details for a specific domain.\n\nRequires `domains:read` permission.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "domainId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Domain details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/Domain" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
    "/v1/domains/{domainId}/inboxes": {
      "get": {
        "tags": ["Inboxes"],
        "operationId": "listDomainInboxes",
        "summary": "List inboxes for a domain",
        "description": "Returns all inboxes belonging to a specific domain.\n\nRequires `inboxes:read` permission.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "domainId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of inboxes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Inbox" }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "post": {
        "tags": ["Inboxes"],
        "operationId": "createDomainInbox",
        "summary": "Create an inbox on a domain",
        "description": "Create a new inbox on a specific domain.\n\nRequires `inboxes:write` permission.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "domainId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["local_part"],
                "properties": {
                  "local_part": { "type": "string" },
                  "display_name": { "type": "string" },
                  "webhook": { "$ref": "#/components/schemas/InboxWebhook" },
                  "status": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Inbox created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/Inbox" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/v1/domains/{domainId}/inboxes/{inboxId}": {
      "get": {
        "tags": ["Inboxes"],
        "operationId": "getInbox",
        "summary": "Get an inbox",
        "description": "Returns a single inbox by domain and inbox ID.\n\nRequires `inboxes:read` permission.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "domainId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "inboxId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Inbox",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/Inbox" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "put": {
        "tags": ["Inboxes"],
        "operationId": "updateInbox",
        "summary": "Update an inbox",
        "description": "Update an inbox's local part, webhook configuration, or status.\n\nRequires `inboxes:write` permission.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "domainId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "inboxId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "local_part": { "type": "string" },
                  "webhook": { "$ref": "#/components/schemas/InboxWebhook" },
                  "status": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated inbox",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/Inbox" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "delete": {
        "tags": ["Inboxes"],
        "operationId": "deleteInbox",
        "summary": "Delete an inbox",
        "description": "Permanently delete an inbox. This does not delete stored messages.\n\nRequires `inboxes:write` permission.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "domainId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "inboxId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Inbox deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "ok": { "type": "boolean", "example": true }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
  },
...
}